chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/querySelector-with-detached-node-distribution-expected.txt

crbug.com/337618: querySelector needs ensure distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0].id should be bar. Threw exception TypeError: Cannot read property 'id' of undefined
PASS barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0] is undefined
PASS successfullyParsed is true

TEST COMPLETE
